destination-service/requirements.txt
Flask
PyJWT
flask-restx